+++ react.spec.js
@@ -1,0 +1,5 @@
+/* eslint-env jest */
+import * as lib from './react'
+import tests from './misc/test-suite.react'
+
+tests(lib)
